sistema_backend/requirements.txt
Django==4.2.22
djangorestframework==3.16.1
django-filter>=24,<26
django-cors-headers==4.7.0
pymysql==1.1.1
python-dateutil==2.9.0.post0
requests==2.32.3
cryptography==42.0.8
bcrypt==4.2.0
